Fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@40160 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil_curl.cpp
@@ -1089,7 +1089,8 @@ retry:
                 }
             }
         }
-        else if ( IsDirectoryFromExists(osVerb, response_code) )
+        else if ( IsDirectoryFromExists(osVerb,
+                                        static_cast<int>(response_code)) )
         {
             eExists = EXIST_YES;
             fileSize = 0;
